services/s18102_image_editor/pyproject.toml
[project]
name = "image-editor-orchestrator"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.28",
    "Pillow>=10.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
exclude = ["demos*", "tests*", "configs*"]

[project.optional-dependencies]
test = ["pytest>=7.0", "matplotlib>=3.7"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
